requirements.txt
Flask
transformers
torch
httpx
librosa
SpeechRecognition
Flask-Session
Flask-SQLAlchemy
python-dotenv

